move PeerID to hyperflux
