+++ packages/client-core/src/components/UserMediaWindows/index.tsx
@@ -26,10 +26,9 @@ Ethereal Engine. All Rights Reserved.
 import { useHookstate } from '@hookstate/core'
 import React from 'react'
 
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { UserID } from '@etherealengine/common/src/schema.type.module'
 import { Engine } from '@etherealengine/ecs/src/Engine'
-import { getMutableState } from '@etherealengine/hyperflux'
+import { PeerID, getMutableState } from '@etherealengine/hyperflux'
 
 import { NetworkState } from '@etherealengine/network'
 import { useMediaNetwork } from '../../common/services/MediaInstanceConnectionService'
