+++ packages/client-core/src/media/PeerMedia.tsx
@@ -25,9 +25,8 @@ Ethereal Engine. All Rights Reserved.
 
 import React, { useEffect } from 'react'
 
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { Engine } from '@etherealengine/ecs/src/Engine'
-import { dispatchAction, getMutableState, getState, useHookstate } from '@etherealengine/hyperflux'
+import { PeerID, dispatchAction, getMutableState, getState, useHookstate } from '@etherealengine/hyperflux'
 import {
   NetworkState,
   screenshareAudioDataChannelType,
