+++ packages/engine/src/interaction/systems/GrabbableSystem.test.ts
@@ -27,10 +27,15 @@ import assert, { strictEqual } from 'assert'
 import { Quaternion, Vector3 } from 'three'
 
 import { NetworkId } from '@etherealengine/common/src/interfaces/NetworkId'
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { AvatarID, UserID, UserName } from '@etherealengine/common/src/schema.type.module'
 import { EntityUUID } from '@etherealengine/ecs'
-import { applyIncomingActions, clearOutgoingActions, dispatchAction, getMutableState } from '@etherealengine/hyperflux'
+import {
+  PeerID,
+  applyIncomingActions,
+  clearOutgoingActions,
+  dispatchAction,
+  getMutableState
+} from '@etherealengine/hyperflux'
 
 import { getComponent, hasComponent, removeComponent, setComponent } from '@etherealengine/ecs/src/ComponentFunctions'
 import { Engine, destroyEngine } from '@etherealengine/ecs/src/Engine'
