+++ packages/network/src/NetworkObjectComponent.ts
@@ -24,7 +24,6 @@ Ethereal Engine. All Rights Reserved.
 */
 
 import { NetworkId } from '@etherealengine/common/src/interfaces/NetworkId'
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { UserID } from '@etherealengine/common/src/schema.type.module'
 import ECS, {
   Component,
@@ -40,7 +39,7 @@ import ECS, {
   useComponent,
   useEntityContext
 } from '@etherealengine/ecs'
-import { Validator, matches } from '@etherealengine/hyperflux'
+import { PeerID, Validator, matches } from '@etherealengine/hyperflux'
 import type { ISchema } from 'bitecs'
 import { useLayoutEffect } from 'react'
