+++ packages/network/src/NetworkState.ts
@@ -23,9 +23,9 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { PeerID, PeersUpdateType } from '@etherealengine/common/src/interfaces/PeerID'
 import { ChannelID, InstanceID, LocationID, RoomCode, UserID } from '@etherealengine/common/src/schema.type.module'
 import {
+  PeerID,
   Validator,
   defineAction,
   defineState,
@@ -38,6 +38,13 @@ import { Spark } from 'primus'
 import { DataChannelType } from './DataChannelRegistry'
 import { Network } from './Network'
 import { SerializationSchema } from './serialization/Utils'
+
+export type PeersUpdateType = {
+  peerID: PeerID
+  peerIndex: number
+  userID: UserID
+  userIndex: number
+}
 
 export interface NetworkPeer {
   userId: UserID
