+++ packages/network/src/functions/NetworkPeerFunctions.test.tsx
@@ -26,10 +26,9 @@ Ethereal Engine. All Rights Reserved.
 import assert from 'assert'
 
 import { NetworkId } from '@etherealengine/common/src/interfaces/NetworkId'
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { InstanceID, UserID } from '@etherealengine/common/src/schema.type.module'
 import { EntityUUID } from '@etherealengine/ecs'
-import { applyIncomingActions, dispatchAction, getMutableState } from '@etherealengine/hyperflux'
+import { PeerID, applyIncomingActions, dispatchAction, getMutableState } from '@etherealengine/hyperflux'
 
 import { Engine, destroyEngine } from '@etherealengine/ecs/src/Engine'
 import { createEngine } from '@etherealengine/spatial/src/initializeEngine'
