+++ packages/network/src/functions/NetworkPeerFunctions.ts
@@ -23,9 +23,8 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { UserID } from '@etherealengine/common/src/schema.type.module'
-import { getMutableState, none } from '@etherealengine/hyperflux'
+import { PeerID, getMutableState, none } from '@etherealengine/hyperflux'
 import { Action } from '@etherealengine/hyperflux/functions/ActionFunctions'
 
 import { Engine } from '@etherealengine/ecs/src/Engine'
