+++ packages/network/src/serialization/DataReader.test.ts
@@ -27,9 +27,8 @@ import assert, { strictEqual } from 'assert'
 import { TypedArray } from 'bitecs'
 
 import { NetworkId } from '@etherealengine/common/src/interfaces/NetworkId'
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { UserID } from '@etherealengine/common/src/schema.type.module'
-import { getMutableState, getState } from '@etherealengine/hyperflux'
+import { PeerID, getMutableState, getState } from '@etherealengine/hyperflux'
 
 import { getComponent, removeComponent, setComponent } from '@etherealengine/ecs/src/ComponentFunctions'
 import { ECSState } from '@etherealengine/ecs/src/ECSState'
