+++ packages/network/src/transports/mediasoup/MediasoupMediaProducerConsumerState.tsx
@@ -23,11 +23,11 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { ChannelID, InstanceID } from '@etherealengine/common/src/schema.type.module'
 import { isClient } from '@etherealengine/common/src/utils/getEnvironment'
 import {
   NO_PROXY_STEALTH,
+  PeerID,
   Validator,
   defineAction,
   defineState,
