+++ packages/ui/src/components/Chat/VideoCall.tsx
@@ -39,10 +39,9 @@ import {
   toggleWebcamPaused
 } from '@etherealengine/client-core/src/transports/SocketWebRTCClientFunctions'
 import { useUserAvatarThumbnail } from '@etherealengine/client-core/src/user/functions/useUserAvatarThumbnail'
-import { PeerID } from '@etherealengine/common/src/interfaces/PeerID'
 import { UserName, userPath } from '@etherealengine/common/src/schema.type.module'
 import { Engine } from '@etherealengine/ecs/src/Engine'
-import { State, getMutableState, useHookstate } from '@etherealengine/hyperflux'
+import { PeerID, State, getMutableState, useHookstate } from '@etherealengine/hyperflux'
 import { NetworkState } from '@etherealengine/network'
 import { useGet } from '@etherealengine/spatial/src/common/functions/FeathersHooks'
 import { t } from 'i18next'
